+++ lib/yt_potion/resources/search.ex
@@ -1,0 +1,7 @@
+defmodule YtPotion.Resources.Search do
+  @moduledoc """
+  The YouTube Data API `search` resource.
+  """
+
+  use YtPotion.Resource, path: "search", actions: [:list]
+end